# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.8.0
  description: API PNIGVS — portail public, référentiel établissements, auth et application métier
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activities/public:
    get:
      operationId: listPublicActivities
      tags: [workflow]
      summary: Activités scolaires autorisées — consultation publique
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/ActivityTypeParam"
        - $ref: "#/components/parameters/EstablishmentIdParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
  /activities:
    get:
      operationId: listActivities
      tags: [workflow]
      summary: Liste des activités scolaires
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityListResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createActivity
      tags: [workflow]
      summary: Créer une activité scolaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Activité créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivitySummary"
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /activities/{id}:
    get:
      operationId: getActivityById
      tags: [workflow]
      summary: Détail d'une activité
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Activité trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivitySummary"
        "404":
          description: Activité introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /requests:
    get:
      operationId: listRequests
      tags: [workflow]
      summary: Liste des demandes d'autorisation
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/RequestStatusParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestListResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createRequest
      tags: [workflow]
      summary: Créer une demande (brouillon)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestBody"
      responses:
        "201":
          description: Demande créée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /requests/{id}:
    get:
      operationId: getRequestById
      tags: [workflow]
      summary: Détail d'une demande avec historique
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Demande trouvée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "404":
          description: Demande introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /requests/{id}/transitions:
    post:
      operationId: transitionRequest
      tags: [workflow]
      summary: Transition workflow d'une demande
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestTransitionBody"
      responses:
        "200":
          description: Transition appliquée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestDetail"
        "400":
          description: Transition invalide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Action non autorisée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /documents/public:
    get:
      operationId: listPublicDocuments
      tags: [documents]
      summary: Bibliothèque documentaire publique
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/DocumentCategoryParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
  /app/documents:
    get:
      operationId: listAppDocuments
      tags: [documents]
      summary: Fichiers scolaires du périmètre utilisateur
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - $ref: "#/components/parameters/DocumentCategoryParam"
      responses:
        "200":
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Déposer un fichier scolaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Document enregistré
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentSummary"
        "400":
          description: Paramètres invalides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /documents/{id}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Télécharger un document
      description: |
        Accès public si le document est marqué public, sinon session authentifiée
        avec périmètre compatible requise.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Fichier binaire
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Non authentifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Accès refusé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Document introuvable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    RequestStatusParam:
      name: status
      in: query
      schema:
        type: string
        enum: [draft, submitted, under_review, forwarded, approved, rejected, cancelled, archived, returned_for_correction]
    DocumentCategoryParam:
      name: category
      in: query
      schema:
        $ref: "#/components/schemas/DocumentCategory"
...
    ActivityTypeParam:
      name: activityType
      in: query
      description: Type d'activité scolaire (sortie, compétition, etc.)
      schema:
        type: string
        maxLength: 64
...
    EstablishmentIdParam:
      name: establishmentId
      in: query
      description: Filtrer les publications d'un établissement (UUID)
      schema:
        type: string
        format: uuid
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required:
            - code
            - message
      required:
        - error
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - pageSize
        - total
        - totalPages
...
    ActivitySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        establishmentId:
          type: string
          format: uuid
        establishmentName:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        location:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - establishmentId
        - establishmentName
        - type
        - title
        - createdAt
    ActivityListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ActivitySummary"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - pagination
    CreateActivityBody:
      type: object
      properties:
        establishmentId:
          type: string
          format: uuid
        type:
          type: string
          maxLength: 64
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        scheduledAt:
          type: string
          format: date-time
        location:
          type: string
          maxLength: 500
      required:
        - type
        - title
    RequestStatus:
      type: string
      enum: [draft, submitted, under_review, forwarded, approved, rejected, cancelled, archived, returned_for_correction]
    WorkflowAction:
      type: string
      enum: [submit, review, forward, approve, publish, reject, cancel, return_for_correction, resubmit, revoke]
    RequestSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RequestStatus"
        description:
          type: string
          nullable: true
        decisionReason:
          type: string
          nullable: true
        establishmentId:
          type: string
          format: uuid
        establishmentName:
          type: string
        activityId:
          type: string
          format: uuid
          nullable: true
        activityTitle:
          type: string
          nullable: true
        checklist:
          type: object
          additionalProperties:
            type: boolean
          description: Checklist établissement (Voyage Découverte, formulaire DVS)
        dvsValidation:
          type: object
          additionalProperties:
            type: boolean
          description: Contrôles internes DVS (prospection, TDR, certificat)
        submittedAt:
          type: string
          format: date-time
          nullable: true
        decidedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - establishmentId
        - establishmentName
        - createdAt
        - updatedAt
    RequestHistoryEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        previousStatus:
          allOf:
            - $ref: "#/components/schemas/RequestStatus"
          nullable: true
        newStatus:
          $ref: "#/components/schemas/RequestStatus"
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - newStatus
        - createdAt
    RequestDetail:
      allOf:
        - $ref: "#/components/schemas/RequestSummary"
        - type: object
          properties:
            history:
              type: array
              items:
                $ref: "#/components/schemas/RequestHistoryEntry"
            allowedActions:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowAction"
          required:
            - history
            - allowedActions
    RequestListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RequestSummary"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - pagination
    CreateRequestBody:
      type: object
      properties:
        activityId:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 5000
      required:
        - activityId
    RequestTransitionBody:
      type: object
      properties:
        action:
          $ref: "#/components/schemas/WorkflowAction"
        reason:
          type: string
          maxLength: 2000
        checklist:
          type: object
          additionalProperties:
            type: boolean
          description: Mise à jour checklist établissement (soumission / resoumission)
        dvsValidation:
          type: object
          additionalProperties:
            type: boolean
          description: Mise à jour contrôles DVS (validation)
      required:
        - action
...
    DocumentCategory:
      type: string
      enum: [circulaire, rapport, fichier_scolaire, autre]
    DocumentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          $ref: "#/components/schemas/DocumentCategory"
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          nullable: true
        isPublic:
          type: boolean
        downloadUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - fileName
        - mimeType
        - isPublic
        - downloadUrl
        - createdAt
    DocumentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSummary"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - pagination
    CreateDocumentBody:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        category:
          $ref: "#/components/schemas/DocumentCategory"
        isPublic:
          type: boolean
          default: false
        fileName:
          type: string
          maxLength: 255
        mimeType:
          type: string
          maxLength: 128
        fileContentBase64:
          type: string
          maxLength: 7000000
      required:
        - title
        - category
        - fileName
        - mimeType
        - fileContentBase64